Move hoogleCmd from Main to Stack.Hoogle
